requirements_fastapi.txt
FastAPI
uvicorn
torch
torchvision
Pillow
pydantic
python-multipart
numpy
scikit-learn